Backend/requirement.txt
fastapi
uvicorn
pydantic
transformers
torch
Pillow
python-dotenv
